Fix CI build
Upgrade dotty to 3.3.0 to match other zio packages.

Use `++version` instead of `++verison!` in the CI,
because the latter syntax forces a Scala version for all
subprojects ignoring the compatible versions declared in crossScalaVersions.

For example; `++3.3.0!` would force the sbt plugin to compile with
Scala version 3.3.0, while ++3.3.0 would recognize that the project
is incompatible with dotty and skip it.

It is crucial to only build the sbt plugin with Scala version 2.12,
which is what sbt 1 itself is built with.