OAuth2: Add support for refresh tokens on Github (#223)
Github apps have optional expiration of access tokens, which leads to
using refresh token to obtain new access token (and new refresh
token). This patch adds support for refresh tokens on Github, which
was previously missing.

Although the Github documentation specifies that 'client_secret' is a
required parameter for refreshing token, it is not required in this
case for unknown reason (perhaps device flow?).

Additionally after this change, new refresh token will be used
preferentially, if present in refresh token response.